App/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
bcrypt==4.1.2
email-validator==2.1.0.post1
psycopg2-binary==2.9.9